feat(pegboard): add logs from edge
