+++ packages/infra/pegboard/container-runner/Cargo.toml
@@ -7,6 +7,7 @@ license = "Apache-2.0"
 
 [dependencies]
 anyhow = "1.0.79"
+nix = { version = "0.27", default-features = false, features = ["fs"] }
 serde = { version = "1.0.195", features = ["derive"] }
 serde_json = "1.0.111"
 signal-hook = "0.3.17"
@@ -15,4 +16,3 @@ signal-hook = "0.3.17"
 portpicker = "0.1.1"
 tempfile = "3.9.0"
 uuid = { version = "1.6.1", features = ["v4"] }
-
