+++ packages/infra/pegboard/manager/Cargo.toml
@@ -15,7 +15,7 @@ futures-util = { version = "0.3" }
 hyper = { version = "0.14", features = ["server", "http1", "tcp"] }
 indoc = "2.0"
 lazy_static = "1.4"
-nix = { version = "0.27", default-features = false, features = ["user", "signal"] }
+nix = { version = "0.27", default-features = false, features = ["fs", "user", "signal"] }
 notify = { version = "6.1.1", default-features = false, features = ["serde", "fsevent-sys"] }
 prometheus = "0.13"
 rand = "0.8"
